ShapeReader modifies debugging input.
The code needs to be able got handle reversed segments, but the "simple" input dataset doesn't contain any. This patch checks for the presence of certain IDs found in the "simple" dataset and reverses input lines with theses IDs. (Collateral damage to other datasets might be possible, but is unlikely because of "simple"'s use of negative way IDs.)