Fixes CorrelationEdges crossing each other.
The original design called for a restriction of all correlated points to be of the same category (start/end) as the origin point. At this point, it isn't clear to me why I considered this restriction in the first place. Removing it and instead considering all points of all parallel segment (which is what this patch does) seems to fix the issue.