Fixes the QGIS Shapefile Snapping issue.
When editing a Shapefile, QGIS 1.8.0 sometimes snaps vertices of linestrings to coordinates that don't quite match those of the vertix they're being snapped to. This appears to be a bug in QGIS. The mismatch distance seems to be very small, usually just in the least significant bit of the double's significand. This fix uses a very small epsilon for node comparison, which appears to get rid of this issue.