Fixes "forward/backward skip" issue #111.
<http://sw.thaw.de/bugs/view.php?id=111>
The earlier fix was dependant upon independent characteristics of the algorithm. In particular, it required the graph to contain zero-length edges at merge points, which additionally needed to be present in the right order based on the input data. This fix simply prevents generalisation to be _started_ at nodes with more than 2 segments connected to it.