Store connectors from end of generalised line to the original line network.
On first small step towards keeping the topology intact: store connectors from the beginning and end of a combined line to the remaining data. This is still buggy and in its current form not all that useful. For example, these connectors are stored even where the existing data is completely removed from the output (like because the two parallels end there without any continuation whatsoever, or because the continuation has also been fully generalised).