Concatenate ungeneralised segments to normalised linestrings.
Those LineSegments that are not combined to a GeneralisedSection due to their parallelism have so far not been integrated into the result, forcing the client to manually sift through the entire collection of input segments and pass through those that weren't generalised. This only increases the size of the output data. This patch introduces a new class provisionally named Section, instances of which represent a single normalised linestring ready for inclusion in the result.