Unified interface for generalised and ungeneralised sections.
Because all result line sections end up in the same Shapefile, which is written by a single method in the Output class, it's helpful to be able to address all sections the same. The SectionInterface (provisional name) enables us to do just that.

Also, includes a fix for the stats gathering code in Combiner.