Incorporate OSM tags into the output data.
Obviously, attributes are required to interpret the result in a meaningful way. This patch not only passes through the highway and ref tags where possible, but also attempts to find applicable values for merged and combined lines. For example, a long road (single carriageway) made out of several ways with the same ref value are merged to a single result line with that same ref, while conflicting refs cause the result ref value to be empty.