Improved output visualisation.
So far the output visualisation in QGIS was done with a mere collection of simple lines of various colors. This patch adds a new style reminiscent of a classic road map. In particular, this new style has a special signature for dual carriageways as opposed to single carriageways.
Also, fixes a regression in data/testbed-nrw-prepare.sh