MutableIterator2: a mostly futile attempt to improve the concept
The MutableIterator2 collects Collections instead of individual items and acts as an Iterator and an Iterable over the collected Collections. This works well for adding collections, but it's not more efficient for adding single items, as is often necessary in the splitting algorithm. And it doesn't really work that well for removing items -- what we'd need for that is our own List implementation to be able to operate on the Entry objects directly. This fact makes this particular commit for the most part useless.