Refactoring Vector.
The former Vector class was moved into the util package as SimpleVector to make room for Vector being worked into the class structure as an interface. The motto is: "Everything is a Vector."
All LineParts and all CorrelationEdges consist of exactly two points. Instead of using composition by making each of those instances _have_ a SimpleVector, they _are_ now Vector implementors themselves and inherit from an abstract superclass AbstractVector. This measure makes for slightly cleaner code and is expected to decrease memory costs.