clarify addNode interface in OsmDataset
From now on there is only one way to add to the dataset-global nodes store: The getNode method, which returns the canonical instance. The getNodeAt method continues to exist, but simply calls the getNode method. This change encourages the use of the canonical instance instead of relying on the added node actually being newly added to the store.
If properly used in this way, it should now be guaranteed that no two distinct structural instances of OsmNode exist having equal coordinates. (Since OsmNode instances can still be created by the OsmNode constructor, proper use is crucial.)
NB: The new code of the splitAt method will fail when splitting parts of lines without an associated dataset.