converted OsmDataset into an interface
OsmDataset is now an interface that doesn't offer any methods to make (shallow) modifications to the data it provides access to. Therefore, the "dataComplete" checks have become superfluous and were removed.
At this point, the only implementing class is InputDataset, which is a concrete public implementation. It might be useful to change this to match a Builder pattern later on.