Remaining uncommitted changes 2013/2014
First commit after a very long break. Totally unchecked! ... I think some of this stuff had not been committed for a reason, but I don't remember it off the top of my head and am not in a good position to investigate right now. This commit should however clear the slate so that I can focus on what's important now.
NB: It might become necessary to revisit this commit later on.