Cleanup chapter 2
- made figure captions ragged rather than fully justified
- fixed double-label issue in chapter 2
- updated BibTeX bibliography as used in chapter 2
Still several issues with bibliography: Export from Papers is buggy, e. g. web pages aren't exported at all and lots of other information that may or may not be important is missing from the BibTeX file. Also, loads of references in the earlier parts of chapter 2 must be re-checked, and many references aren't even well specified in the TeX source code as of yet.