Final document and subdir structure
The subfiles package enables keeping the individual chapters and their parts separate from the main document file while still allowing everything to be compiled separately without changing any paths in the source. Crucially, for this to work, the main document file needs to be on the same subdirectory level as the chapters are.
This approach is inspired by https://en.wikibooks.org/wiki/LaTeX/Modular_Documents#Subfiles .
Additionally, this commit moves the early document outline to the new main file subdir to keep the project directory clean.