Chapter 4
- new algorithm for CorrelationGraph
- fix "nodes of segment" specification

However, TA noted rightly that this level of formality may already be too detailed, resulting in an unnecessary strain for the reader. The goal should be to use normal language as much as possible. Neither matters of course nor any special cases solved in code should be formally defined. The latter must be mentioned, but the precise details might be left to the implementation, since that, too, is a part of the thesis in a way.