Chapter 4: Formal description complete
The first draft of the algorithms' formal description is now more or less complete. Important details are still missing in places, but these places are now marked accordingly. Filling in the blanks should now be relatively straightforward.

An issue that hasn't been solved are the changes in writing style throughout this chapter: Some algorithms are described using mathematic formulas, others are described in prose; some are described in an imperative style, others use a descriptive style that even in prose reads more like a mathematical definition than an implementable algorithm.

The generalisation algorithm in section 4.3.4 is particularly problematic in that its writing style is informal while at the same time several important edge cases haven't been described at all. The current plan is to do so in prose only for simplicity's sake. Whether that's sufficient remains to be seen.

Also still missing is a coherent narrative tying together all parts of this chapter. This is the next important step.