Solve GeoTools classpath issues
GeoTools tends to complain if certain classes are on the classpath, so the not-that-unusual method of simply putting *everything* on the classpath doesn't work here. This is apparently a well-known issue that the developers refuse to fix or properly document. Since updating GeoTools from version 9 to 17, the EPSG JARs are no longer the only issue where this fails.

The new solution is to specify the exact GeoTools JAR files we use in the java command. This is not very readable and also hard to maintain, but it should make getting this app up and running significantly easier for users and is also easier to document and explain.

An alternative solution (and indeed the one apparently expected by the GeoTools developers) might be to switch from Ant to Maven. Seems like a good idea in the abstract, but it seems smarter to keep things as simple as possible for this particular project for now.

While unrelated to the classpath issues, another fix included here concerns error messages that occur when reading Shapefiles with GeoTools. It used to be that properly closing the SimpleFeatureIterator caused error messages (yes, really). Now, however, *not* closing that iterator will cause error messages. So now we properly close and dispose of the Shapefile resources.

The result is that no more irritating error messages are written by GeoTools, and future versions of GeoTools are expected to just work (provided one adjusts the java classpath in run-combiner.sh).