Update Shapefile output for GeoTools 17
Somewhere between GeoTools 8 and GeoTools 17, the magic constant used to identify the geometry column in a Shapefile changed from "geometry" to "the_geom" for whatever reason. This commit removes this magic constant from this project, replacing it with code fetching its current value from the GeoTools Shapefile data store.

Obviously, the use of such magic constants is a bad idea in the first place, so part of the blame for the Shapefile output no longer working with GeoTools 17 may seem to rest on me. Yet, bizarrely, the official Shapefile example in the Geotools documentation continues to use this magic constant with no words of warning or explanation. At least the official example currently uses the correct magic value -- the API documentation actually suggests one that no longer works. While such troubles are of course not unheard of in an open source project, I'm sorry to say this paints a certain picture of the state of the GeoTools project when viewed together with others of its issues such as the one regarding the classpath mentioned in another commit message on this project a few days ago.