Refactor AbstractSegment for Composite pattern
The Composite pattern seems like a good fit for our fragments of segments that are segments themselves. The core of this Composite implementation is a new FragmentIterator that walks the composition tree and returns every Leaf.

The children of the tree nodes are now implemented as 2-item arrays, which is why SplitQueueListener now accepts a List of fragments rather than two individual fragments, hopefully making this a bit more efficient.

Or perhaps I should say "less inefficient", as initial testing reveals that my test dataset costs about 10 % more CPU time to process. However, since the code is now better structured and the real time required for processing doesn't seem to change at all, I'll just leave it at that.

This change also makes Segments iterable, returning all of their fragments.