Make CorrelationEdge midPoint independent of global node interning
During generalisation most CorrelationEdges are queried for their respective mid points multiple times. It is expected that for each edge the same identical Node object is used by the client each time they are requested. This is currently the case, but only because the CorrelationEdge's clients use the global dataset node interning system. This patch introduces a local cache of the mid point in each CorrelationEdge, which should obviate the need for adding these points to the dataset's node interning for correctness.