Big Renaming (2)
To better match the terms used in the thesis, the class OsmNode is split into Node (interface), AbstractNode (default implementation for most of the interface), concrete SourceNode (nodes with support for our correlation graph), and Nodes (static utility methods). This separation of concerns also improves encapsulation.

Additionally there are two more concrete Node implementations. Nodes.NonexistantNode is used for nodes created during the splitting process. These do not exist in the source dataset and are only used to define fragments for analysis and are discarded afterwards. GeneralisedNodes don't exist in the source dataset either. These are created during generalisation of two parallel segments and define the combined line string.

Note that the AbstractSegment class expects to work with the type Node, while the concrete SourceSegment class overrides those methods to work with SourceNode.

Some parts of the classes that needed to be changed as a result of splitting up OsmNode receive minor improvements on encapsulation as well.