Chapter 4: Major rework
Chapter 4.3.4 is basically completely rewritten. Its algorithm has a new structure and is now properly illustrated. Additionally there are many major and minor changes throughout this chapter. These include, but are not limited to, redefinitions of math symbols, replaced figures, new and renamed algorithms and fixes of logical issues with the text. While this chapter could still be improved, it is now finally at a stage where submission might be tenable.