Add foot note to explain text/figure mismatch in 2.3.2
The point that residential streets are drawn to the same width in osm-carto at zoom 13 and 14 doesn't hold much merit. They were different before 2015, and they were again made different in late 2015 and have been since. This point should probably have been removed entirely, as it's not central to the argument. Seeing as there currently is a mismatch between the text and the figure, we should at least add a footnote to explain that.