Optimise numbering in single-chapter mode
Since the exact count of figures (and tables) in each chapter is now known, we can set the counters manually for each chapter such that the numbering in single-chapter mode exactly matches that of the final main document. Doing this is useful because numbering takes more horizontal space for counts larger than 9, which influences wrapping.

Because BibLaTeX is not usually run in single-chapter mode, the same is not true for page numbers. To avoid any possible misunderstanding, these are changed to roman numerals for single-chapter mode.